requirements.txt
viam-sdk
onnx2torch
numpy
torch
torchvision
facenet-pytorch
opencv-python
pillow
